catalog/backend/requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.0.1
python-decouple==3.8
